+++ src/MinFraud/Model/CreditCard.php
@@ -6,28 +6,29 @@ namespace MaxMind\MinFraud\Model;
  * Class CreditCard
  * @package MaxMind\MinFraud\Model
  *
- * @property string $brand The card brand, such as "Visa", "Discover",
+ * @property string|null $brand The card brand, such as "Visa", "Discover",
  * "American Express", etc.
  *
- * @property string $country This property contains the the two letter {@link
- * http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2 ISO 3166-1 alpha-2 country
- * code} associated with the location of the majority of customers using this
- * credit card as determined by their billing address. In cases where the
- * location of customers is highly mixed, this defaults to the country of the
- * bank issuing the card.
+ * @property string|null $country This property contains the the two letter
+ * {@link http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2 ISO 3166-1 alpha-2
+ * country code} associated with the location of the majority of customers
+ * using this credit card as determined by their billing address. In cases
+ * where the location of customers is highly mixed, this defaults to the
+ * country of the bank issuing the card.
  *
- * @property boolean $isIssuedInBillingAddressCountry This property is true if
- * the country of the billing address matches the country of the majority of
- * customers using this credit card. In cases where the location of customers
- * is highly mixed, the match is to the country of the bank issuing the card.
+ * @property boolean|null $isIssuedInBillingAddressCountry This property is
+ * true if the country of the billing address matches the country of the
+ * majority of customers using this credit card. In cases where the location
+ * of customers is highly mixed, the match is to the country of the bank
+ * issuing the card.
  *
- * @property boolean $isPrepaid This property is true if the card is a prepaid
- * card.
+ * @property boolean|null $isPrepaid This property is true if the card is a
+ * prepaid card.
  *
  * @property \MaxMind\MinFraud\Model\Issuer $issuer An object containing
  * information about the credit card issuer.
  *
- * @property string $type The card's type. The valid values are: charge,
+ * @property string|null $type The card's type. The valid values are: charge,
  * credit, debit.
  */
 class CreditCard extends AbstractModel