+++ src/MinFraud/Model/ShippingAddress.php
@@ -8,12 +8,13 @@ namespace MaxMind\MinFraud\Model;
  *
  * This model contains properties of the shipping address.
  *
- * @property integer $distanceToBillingAddress The distance in kilometers from
- * the shipping address to billing address.
- * @property boolean $isHighRisk This property is true if the shipping address
- * is in the IP country. The property is false when the address is not in the
- * IP country. If the shipping address could not be parsed or was not provided
- * or the IP address could not be geolocated, then the property is null.
+ * @property integer|null $distanceToBillingAddress The distance in kilometers
+ * from the shipping address to billing address.
+ * @property boolean|null $isHighRisk This property is true if the shipping
+ * address is in the IP country. The property is false when the address is not
+ * in the IP country. If the shipping address could not be parsed or was not
+ * provided or the IP address could not be geolocated, then the property is
+ * null.
  */
 class ShippingAddress extends Address
 {